feat(enemy): add enemy hit/attack/seen sounds
