--- Assets/Scripts/Enemy AI/EnemyFootsteps.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: d75350ee09354cd284c9c5919e52ee0f
-timeCreated: 1704216271
