+++ Assets/Sound Effects/Enemies.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5016fec8e278bce4090a2ba5de806167
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
